refactored code for firestore
Implemented firestore database instead of realtime database and refactored user classes to work with that. Created custom userIDs that work with a counter database in firestore to generate an ID(for example, the first user gets "user1" as their ID and this their document name, and so forth). Completely redone all classes:

Added hash map to user class to map user info to firestore database properly

UserDBConnector now connects to the firestore cloud database

UserDB is incomplete at the moment but addUser and updateUser now use the unique userID generated and used as the document name for each user in firestore as part of the document directory, which I believe makes for an easier time navigating our user database