Changed userID usage
Changed the implementation of how we use userID. Now userDB uses the randomly generated UUID from MainActivity that is assigned to a user at first launch. There is now a getter method in MainActivity that gets this userID striing from MainActivity and can  we used in UserDB now that UserDB inherits from MainActivity. We get the userID using this new getter (getUserIDForUserDB) and assign it to a string userID in the addNewUser method. we then set this as the userID of an instance of the User class (user) using the setter method found in the User class.